fix(read_map_utils.c and texture.c): fix comparison between different type
int <-> size_t

+++ src/read_map_utils.c
@@ -53,7 +53,7 @@ void		get_resolution(t_map *map, char *line, int i, int *obj_num)
 		map->window_width = 1920;
 	if (map->window_height > 1080)
 		map->window_height = 1080;
-	if (ft_strlen(line + i) != num_of_spaces(line, i))
+	if (ft_strlen(line + i) != (size_t)num_of_spaces(line, i))
 		cub_file_err();
 	(*obj_num)++;
 }
